% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altair.package.R
\docType{package}
\name{altair}
\alias{altair}
\alias{altair-package}
\title{Altaïr: un paquet pour analyser les bases de paye XML.}
\description{
Altaïr comprend les fonctions suivantes:
effectifs, bar and baz.
}
\section{démographie}{

effectifs : tableau des effectifs, ETP et EQTP
}

