% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectifs.R
\name{pyramide_ages}
\alias{pyramide_ages}
\title{Pyramide des âges.}
\usage{
pyramide_ages(Avant, Après, titre = "",
  date.début = début.période.sous.revue, date.fin = fin.période.sous.revue,
  couleur_H = "darkslateblue", couleur_F = "firebrick4")
}
\arguments{
\item{Avant}{data.table décrivant la situation en début de période
Cette base doit avoir la forme suivante : 
\tabular{ccc}{
  age \tab Hommes \tab Femmes \cr
  15  \tab   0  \tab    1   \cr
  16  \tab   NA \tab    2   \cr
  17  \tab   1  \tab    3   \cr
  18  \tab  409 \tab    52  \cr
  ... \tab  ... \tab ...    \cr    
  68  \tab 2216 \tab    NA  
}}

\item{Après}{data.table décrivant la situation en fin de période. Même format que Avant.}

\item{titre}{Titre du graphique}

\item{date.début}{date du début de la période}

\item{date.fin}{date de fin de période}

\item{couleur_H}{couleur utilisée pour représenter les hommes (partie gauche de la pyramide). Par défaut \code{darkslateblue}}

\item{couleur_F}{couleur utilisée pour représenter les femmes (partie droite de la pyramide). Par défaut \code{firebrick4}}
}
\value{
Un graphique comprenat une pyramide, une légende et éventuellement un titre.
}
\description{
Elabore une pyramide des âges verticale avec superposition du début et de la fin de la période sous revue.
}
\examples{
pyramide_ages(df1, df2, "Pyramide des âges", 2008, 2012)
}

