% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectifs.R
\name{effectifs}
\alias{effectifs}
\title{Tableau des effectifs.}
\usage{
effectifs(période, Bulletins = Bulletins.paie,
  personnels = Analyse.rémunérations,
  Analyse.vpe = Analyse.variations.par.exercice,
  Analyse.v = Analyse.variations)
}
\arguments{
\item{période}{Vecteur des années de la période sous revue}

\item{Bulletins}{Base des bulletins de paye}

\item{Analyse}{Base des analyses de rémunérations}

\item{Analyse}{Base des analyses de variations de rémunérations}
}
\value{
Un tableau des effectifs mis en forme.
}
\description{
Elabore un tableau des effectifs et équivalents temps plein travaillés par type de personnel et par année.
}
\examples{
effectifs(2010:2015)
}

